% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteBlosum.R
\name{DiscreteBlosum}
\alias{DiscreteBlosum}
\title{Discrete blosum descriptor.}
\usage{
DiscreteBlosum(x)
}
\arguments{
\item{x}{A string of amino acid letters}
}
\value{
A 20 dimensional numeric vector
}
\description{
\code{DiscreteBlosum} returns the sum of blosum descriptors of amino acids in
a protein sequence.
}
\examples{
x = "LALHLLLLHMHMMDRSLLLH"
DiscreteBlosum(x)
}

