% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteSequentialSparseFiveParts.R
\name{DiscreteSequentialSparseFiveParts}
\alias{DiscreteSequentialSparseFiveParts}
\title{Discrete sequential sparse descriptor with split number=5.}
\usage{
DiscreteSequentialSparseFiveParts(x)
}
\arguments{
\item{x}{A string of amino acid letters}
}
\value{
A 100 dimensional numeric vector
}
\description{
\code{DiscreteSequentialSparseFiveParts} returns the concatenation of the sum
of sparse descriptors of amino acids in each split of a protein sequence.
}
\examples{
x = "LALHLLLLHMHMMDRSLLLH"
DiscreteSequentialSparseFiveParts(x)
}

