% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteSequentialSparseFourParts.R
\name{DiscreteSequentialSparseFourParts}
\alias{DiscreteSequentialSparseFourParts}
\title{Discrete sequential sparse descriptor with split number=4.}
\usage{
DiscreteSequentialSparseFourParts(x)
}
\arguments{
\item{x}{A string of amino acid letters}
}
\value{
A 80 dimensional numeric vector
}
\description{
\code{DiscreteSequentialSparseFourParts} returns the concatenation of the sum
of sparse descriptors of amino acids in each split of a protein sequence.
}
\examples{
x = "LALHLLLLHMHMMDRSLLLH"
DiscreteSequentialSparseFourParts(x)
}

