% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteSequentialSparseThreeParts.R
\name{DiscreteSequentialSparseThreeParts}
\alias{DiscreteSequentialSparseThreeParts}
\title{Discrete sequential sparse descriptor with split number=3.}
\usage{
DiscreteSequentialSparseThreeParts(x)
}
\arguments{
\item{x}{A string of amino acid letters}
}
\value{
A 60 dimensional numeric vector
}
\description{
\code{DiscreteSequentialSparseThreeParts} returns the concatenation of the
sum of sparse descriptors of amino acids in each split of a protein sequence.
}
\examples{
x = "LALHLLLLHMHMMDRSLLLH"
DiscreteSequentialSparseThreeParts(x)
}

