% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteSequentialSparseTwoParts.R
\name{DiscreteSequentialSparseTwoParts}
\alias{DiscreteSequentialSparseTwoParts}
\title{Discrete sequential sparse descriptor with split number=2.}
\usage{
DiscreteSequentialSparseTwoParts(x)
}
\arguments{
\item{x}{A string of amino acid letters}
}
\value{
A 40 dimensional numeric vector
}
\description{
\code{DiscreteSequentialSparseTwoParts} returns the concatenation of the sum
of sparse descriptors of amino acids in each split of a protein sequence.
}
\examples{
x = "LALHLLLLHMHMMDRSLLLH"
DiscreteSequentialSparseTwoParts(x)
}

