% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteSequentialTdTwoParts.R
\name{DiscreteSequentialTdTwoParts}
\alias{DiscreteSequentialTdTwoParts}
\title{Discrete sequential 3D descriptor with split number=2.}
\usage{
DiscreteSequentialTdTwoParts(x)
}
\arguments{
\item{x}{A string of amino acid letters}
}
\value{
A 6 dimensional numeric vector
}
\description{
\code{DiscreteSequentialTdTwoParts} returns the concatenation of the sum of
3D descriptors of amino acids in each split of a protein sequence.
}
\examples{
x = "LALHLLLLHMHMMDRSLLLH"
DiscreteSequentialTdTwoParts(x)
}

