% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscreteSparse.R
\name{DiscreteSparse}
\alias{DiscreteSparse}
\title{Discrete sparse descriptor.}
\usage{
DiscreteSparse(x)
}
\arguments{
\item{x}{A string of amino acid letters}
}
\value{
A 20 dimensional numeric vector
}
\description{
\code{DiscreteSparse} returns the sum of sparse descriptors of amino acids in
a protein sequence.
}
\examples{
x = "LALHLLLLHMHMMDRSLLLH"
DiscreteSparse(x)
}

