% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LastFirstDiscreteSparse.R
\name{LastFirstDiscreteSparse}
\alias{LastFirstDiscreteSparse}
\title{Last first discrete sparse descriptor.}
\usage{
LastFirstDiscreteSparse(x)
}
\arguments{
\item{x}{A string of amino acid letters}
}
\value{
A 40 dimensional numeric vector
}
\description{
\code{LastFirstDiscreteSparse} returns the concatenation of the sum of sparse
descriptors of amino acids in a protein sequence and the sum of combination
of last part of an amino acid descriptor with the first part of its neighbour
amino acid descriptor.
}
\examples{
x = "LALHLLLLHMHMMDRSLLLH"
LastFirstDiscreteSparse(x)
}

