% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequentialBlosum.R
\name{SequentialBlosum}
\alias{SequentialBlosum}
\title{Sequential blosum descriptor.}
\usage{
SequentialBlosum(x)
}
\arguments{
\item{x}{A string of amino acid letters}
}
\value{
A 20*n dimensional numeric vector where n is the protein length
}
\description{
\code{SequentialBlosum} returns the concatenation of blosum descriptors of
amino acids in a protein sequence.
}
\examples{
x = "LALHLLLLHMHMMDRSLLLH"
SequentialBlosum(x)
}

