% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequentialDiscreteBlosum.R
\name{SequentialDiscreteBlosum}
\alias{SequentialDiscreteBlosum}
\title{Sequential discrete blosum descriptor with step size=2.}
\usage{
SequentialDiscreteBlosum(x)
}
\arguments{
\item{x}{A string of amino acid letters}
}
\value{
A 40 dimensional numeric vector
}
\description{
\code{SequentialDiscreteBlosum} returns the sum of the concatenation of
blosum descriptors of amino acids at every step size in a protein sequence.
}
\examples{
x = "LALHLLLLHMHMMDRSLLLH"
SequentialDiscreteBlosum(x)
}

