% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequentialDiscreteSparse.R
\name{SequentialDiscreteSparse}
\alias{SequentialDiscreteSparse}
\title{Sequential discrete sparse descriptor with step size=2.}
\usage{
SequentialDiscreteSparse(x)
}
\arguments{
\item{x}{A string of amino acid letters}
}
\value{
A 40 dimensional numeric vector
}
\description{
\code{SequentialDiscreteSparse} returns the sum of the concatenation of
sparse descriptors of amino acids at every step size in a protein sequence.
}
\examples{
x = "LALHLLLLHMHMMDRSLLLH"
SequentialDiscreteSparse(x)
}

