% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequentialDiscreteTd.R
\name{SequentialDiscreteTd}
\alias{SequentialDiscreteTd}
\title{Sequential discrete 3D descriptor with step size=2.}
\usage{
SequentialDiscreteTd(x)
}
\arguments{
\item{x}{A string of amino acid letters}
}
\value{
A 6 dimensional numeric vector
}
\description{
\code{SequentialDiscreteTd} returns the sum of the concatenation of 3D
descriptors of amino acids at every step size in a protein sequence.
}
\examples{
x = "LALHLLLLHMHMMDRSLLLH"
SequentialDiscreteTd(x)
}

