% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequentialTd.R
\name{SequentialTd}
\alias{SequentialTd}
\title{Sequential 3D descriptor.}
\usage{
SequentialTd(x)
}
\arguments{
\item{x}{A string of amino acid letters}
}
\value{
A 3*n dimensional numeric vector where n is the protein length
}
\description{
\code{SequentialTd} returns the concatenation of 3D descriptors of amino
acids in a protein sequence.
}
\examples{
x = "LALHLLLLHMHMMDRSLLLH"
SequentialTd(x)
}

