% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiftedDiscreteSparse.R
\name{ShiftedDiscreteSparse}
\alias{ShiftedDiscreteSparse}
\title{Discrete sparse descriptor.}
\usage{
ShiftedDiscreteSparse(x)
}
\arguments{
\item{x}{A string}
}
\value{
A 20 dimensional numeric vector
}
\description{
\code{DiscreteSparse} returns the sum of sparse descriptors of amino acids in a protein
sequence.
}
\examples{
x = "LALHLLLLHMHMMDRSLLLH"
ShiftedDiscreteSparse(x)
}

